% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{analytical_inference}
\alias{analytical_inference}
\title{Analytical inference using influence functions}
\usage{
analytical_inference(effects, diff_estimate, effect_measure, alpha)
}
\arguments{
\item{effects}{Effect estimates}

\item{diff_estimate}{Difference estimate}

\item{effect_measure}{Effect measure}

\item{alpha}{Alpha level}
}
\value{
Analytical inference results
}
\description{
Analytical inference using influence functions
}
