% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{assess_balance_quality}
\alias{assess_balance_quality}
\title{Assess overall balance quality}
\usage{
assess_balance_quality(smd_between, smd_rct, smd_obs)
}
\arguments{
\item{smd_between}{SMD between datasets}

\item{smd_rct}{SMD within RCT}

\item{smd_obs}{SMD within observational}
}
\value{
Balance quality assessment
}
\description{
Assess overall balance quality
}
