% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{assess_fragility}
\alias{assess_fragility}
\title{Assess fragility of results}
\usage{
assess_fragility(diagnostics)
}
\arguments{
\item{diagnostics}{Diagnostic results}
}
\value{
Fragility assessment
}
\description{
Assess fragility of results
}
