% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{assess_method_robustness}
\alias{assess_method_robustness}
\title{Assess estimation method robustness}
\usage{
assess_method_robustness(effects)
}
\arguments{
\item{effects}{Effect estimates}
}
\value{
Method robustness assessment
}
\description{
Assess estimation method robustness
}
