% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{assess_model_robustness}
\alias{assess_model_robustness}
\title{Assess model robustness}
\usage{
assess_model_robustness(model_diagnostics)
}
\arguments{
\item{model_diagnostics}{Model diagnostics}
}
\value{
Model robustness assessment
}
\description{
Assess model robustness
}
