% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{assess_robustness}
\alias{assess_robustness}
\title{Assess robustness of results}
\usage{
assess_robustness(effects, diagnostics)
}
\arguments{
\item{effects}{Effect estimates}

\item{diagnostics}{Diagnostic results}
}
\value{
Robustness assessment
}
\description{
Assess robustness of results
}
