% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{batch_analysis}
\alias{batch_analysis}
\title{Batch analysis for multiple datasets}
\usage{
batch_analysis(rct_data_list, obs_data_list, formula, estimator = "aipw", ...)
}
\arguments{
\item{rct_data_list}{List of RCT datasets}

\item{obs_data_list}{List of observational datasets}

\item{formula}{Model formula}

\item{estimator}{Estimator to use}

\item{...}{Additional arguments passed to unconfoundedness_test}
}
\value{
List of results
}
\description{
Batch analysis for multiple datasets
}
