% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{classify_overlap_quality}
\alias{classify_overlap_quality}
\title{Classify overlap quality}
\usage{
classify_overlap_quality(extreme_rct, extreme_obs, overlap_stat)
}
\arguments{
\item{extreme_rct}{Proportion of extreme PS in RCT}

\item{extreme_obs}{Proportion of extreme PS in observational}

\item{overlap_stat}{Overlap statistic}
}
\value{
Overlap quality classification
}
\description{
Classify overlap quality
}
