% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{coerce_binary}
\alias{coerce_binary}
\title{Coerce variable to binary 0/1}
\usage{
coerce_binary(x, var_name = "Variable")
}
\arguments{
\item{x}{Variable to coerce}

\item{var_name}{Variable name for error messages}
}
\value{
Binary 0/1 vector
}
\description{
Coerce variable to binary 0/1
}
