% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_balance_diagnostics}
\alias{compute_balance_diagnostics}
\title{Compute balance diagnostics}
\usage{
compute_balance_diagnostics(X_rct, X_obs, A_rct, A_obs, w_rct, w_obs)
}
\arguments{
\item{X_rct}{RCT covariates}

\item{X_obs}{Observational covariates}

\item{A_rct}{RCT treatment}

\item{A_obs}{Observational treatment}

\item{w_rct}{RCT weights}

\item{w_obs}{Observational weights}
}
\value{
Balance diagnostics
}
\description{
Compute balance diagnostics
}
