% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_e_value}
\alias{compute_e_value}
\title{Compute E-value for sensitivity to unmeasured confounding}
\usage{
compute_e_value(obs_effect, rct_effect, effect_measure)
}
\arguments{
\item{obs_effect}{Observational effect}

\item{rct_effect}{RCT effect}

\item{effect_measure}{Effect measure}
}
\value{
E-value
}
\description{
Compute E-value for sensitivity to unmeasured confounding
}
