% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_effect_difference}
\alias{compute_effect_difference}
\title{Compute effect difference}
\usage{
compute_effect_difference(rct_effect, obs_effect, effect_measure)
}
\arguments{
\item{rct_effect}{RCT effect estimate}

\item{obs_effect}{Observational effect estimate}

\item{effect_measure}{Effect measure}
}
\value{
Difference estimate
}
\description{
Compute effect difference
}
