% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_inference}
\alias{compute_inference}
\title{Compute inference (confidence intervals and p-values)}
\usage{
compute_inference(
  parsed_data,
  estimator,
  effect_measure,
  family_y,
  weights,
  effects,
  diff_estimate,
  inference_method,
  B,
  alpha,
  parallel,
  n_cores
)
}
\arguments{
\item{parsed_data}{Parsed data}

\item{estimator}{Estimator}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}

\item{weights}{Weight object}

\item{effects}{Effect estimates}

\item{diff_estimate}{Difference estimate}

\item{inference_method}{Inference method}

\item{B}{Bootstrap replicates}

\item{alpha}{Alpha level}

\item{parallel}{Use parallel processing}

\item{n_cores}{Number of cores}
}
\value{
Inference results
}
\description{
Compute inference (confidence intervals and p-values)
}
