% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_model_diagnostics}
\alias{compute_model_diagnostics}
\title{Compute model diagnostics}
\usage{
compute_model_diagnostics(effects, family_y)
}
\arguments{
\item{effects}{Effect estimates}

\item{family_y}{Outcome family}
}
\value{
Model diagnostics
}
\description{
Compute model diagnostics
}
