% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_overlap_diagnostics}
\alias{compute_overlap_diagnostics}
\title{Compute overlap/positivity diagnostics}
\usage{
compute_overlap_diagnostics(A_rct, A_obs, e_rct, e_obs)
}
\arguments{
\item{A_rct}{RCT treatment}

\item{A_obs}{Observational treatment}

\item{e_rct}{RCT propensity scores}

\item{e_obs}{Observational propensity scores}
}
\value{
Overlap diagnostics
}
\description{
Compute overlap/positivity diagnostics
}
