% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_overlap_statistic}
\alias{compute_overlap_statistic}
\title{Compute overlap statistic between two distributions}
\usage{
compute_overlap_statistic(x1, x2)
}
\arguments{
\item{x1}{First distribution}

\item{x2}{Second distribution}
}
\value{
Overlap statistic
}
\description{
Compute overlap statistic between two distributions
}
