% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_sample_diagnostics}
\alias{compute_sample_diagnostics}
\title{Compute sample size diagnostics}
\usage{
compute_sample_diagnostics(parsed_data, w_rct, w_obs)
}
\arguments{
\item{parsed_data}{Parsed data}

\item{w_rct}{RCT weights}

\item{w_obs}{Observational weights}
}
\value{
Sample size diagnostics
}
\description{
Compute sample size diagnostics
}
