% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_sensitivity_analysis}
\alias{compute_sensitivity_analysis}
\title{Compute sensitivity analysis}
\usage{
compute_sensitivity_analysis(effects, diagnostics, effect_measure)
}
\arguments{
\item{effects}{Effect estimates}

\item{diagnostics}{Diagnostic results}

\item{effect_measure}{Effect measure}
}
\value{
Sensitivity analysis results
}
\description{
Compute sensitivity analysis
}
