% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{compute_treatment_balance}
\alias{compute_treatment_balance}
\title{Compute treatment balance within dataset}
\usage{
compute_treatment_balance(X, A, weights)
}
\arguments{
\item{X}{Covariate matrix}

\item{A}{Treatment vector}

\item{weights}{Sample weights}
}
\value{
Treatment balance
}
\description{
Compute treatment balance within dataset
}
