% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{create_diagnostic_summary}
\alias{create_diagnostic_summary}
\title{Create diagnostic summary for summary method}
\usage{
create_diagnostic_summary(diagnostics)
}
\arguments{
\item{diagnostics}{Diagnostics object}
}
\value{
Summarized diagnostics
}
\description{
Create diagnostic summary for summary method
}
