% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{create_latex_table}
\alias{create_latex_table}
\title{Create latex table from results}
\usage{
create_latex_table(x, file = NULL, caption = "Unconfoundedness Test Results")
}
\arguments{
\item{x}{unconf_test object or list of results}

\item{file}{Output file path (NULL for console output)}

\item{caption}{Table caption}
}
\value{
LaTeX table code
}
\description{
Create latex table from results
}
