% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{create_rmd_content}
\alias{create_rmd_content}
\title{Create R Markdown content for diagnostic report}
\usage{
create_rmd_content(x)
}
\arguments{
\item{x}{unconf_test object}
}
\value{
R Markdown content as character vector
}
\description{
Create R Markdown content for diagnostic report
}
