% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimate_matching}
\alias{estimate_matching}
\title{Matching Estimator}
\usage{
estimate_matching(Y, A, X, weights, effect_measure, family_y)
}
\arguments{
\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{X}{Covariate matrix}

\item{weights}{Sample weights}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}
}
\value{
Matching estimate
}
\description{
Matching Estimator
}
