% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{estimate_single_effect}
\alias{estimate_single_effect}
\title{Estimate effect for single dataset}
\usage{
estimate_single_effect(
  Y,
  A,
  X,
  weights,
  estimator,
  effect_measure,
  family_y,
  stabilize,
  trim
)
}
\arguments{
\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{X}{Covariate matrix}

\item{weights}{Sample weights}

\item{estimator}{Estimation method}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}

\item{stabilize}{Stabilize weights}

\item{trim}{Trimming bounds}
}
\value{
Single effect estimate
}
\description{
Estimate effect for single dataset
}
