% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{extract_model_diagnostics}
\alias{extract_model_diagnostics}
\title{Extract diagnostics from GLM model}
\usage{
extract_model_diagnostics(model)
}
\arguments{
\item{model}{GLM model object}
}
\value{
Model diagnostics
}
\description{
Extract diagnostics from GLM model
}
