% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{format_multiple_results_latex}
\alias{format_multiple_results_latex}
\title{Format multiple results for LaTeX}
\usage{
format_multiple_results_latex(results, caption)
}
\arguments{
\item{results}{List of unconf_test objects}

\item{caption}{Table caption}
}
\value{
LaTeX code
}
\description{
Format multiple results for LaTeX
}
