% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{format_p_value_latex}
\alias{format_p_value_latex}
\title{Format p-value for LaTeX}
\usage{
format_p_value_latex(p)
}
\arguments{
\item{p}{P-value}
}
\value{
Formatted p-value
}
\description{
Format p-value for LaTeX
}
