% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{format_single_result_latex}
\alias{format_single_result_latex}
\title{Format single result for LaTeX}
\usage{
format_single_result_latex(x, caption)
}
\arguments{
\item{x}{unconf_test object}

\item{caption}{Table caption}
}
\value{
LaTeX code
}
\description{
Format single result for LaTeX
}
