% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_and_plotting.R
\name{format_summary_table}
\alias{format_summary_table}
\title{Format summary table}
\usage{
format_summary_table(summary_obj)
}
\arguments{
\item{summary_obj}{Summary object}
}
\value{
Formatted data frame
}
\description{
Format summary table
}
