% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_and_plotting.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate recommendations}
\usage{
generate_recommendations(x)
}
\arguments{
\item{x}{unconf_test object}
}
\value{
Character vector of recommendations
}
\description{
Generate recommendations
}
