% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_and_plotting.R
\name{generate_warnings}
\alias{generate_warnings}
\title{Generate warnings based on diagnostics}
\usage{
generate_warnings(x)
}
\arguments{
\item{x}{unconf_test object}
}
\value{
Character vector of warnings
}
\description{
Generate warnings based on diagnostics
}
