% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{interpret_e_value}
\alias{interpret_e_value}
\title{Interpret E-value}
\usage{
interpret_e_value(e_value)
}
\arguments{
\item{e_value}{E-value}
}
\value{
Interpretation
}
\description{
Interpret E-value
}
