% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_and_plotting.R
\name{interpret_test_result}
\alias{interpret_test_result}
\title{Interpret test result}
\usage{
interpret_test_result(p_value, difference)
}
\arguments{
\item{p_value}{P-value from test}

\item{difference}{Estimated difference}
}
\value{
Interpretation string
}
\description{
Interpret test result
}
