% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{parse_formula_and_data}
\alias{parse_formula_and_data}
\title{Parse formula and prepare data matrices}
\usage{
parse_formula_and_data(formula, data_rct, data_obs)
}
\arguments{
\item{formula}{Model formula}

\item{data_rct}{RCT data}

\item{data_obs}{Observational data}
}
\value{
List with parsed components
}
\description{
Parse formula and prepare data matrices
}
