% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{perform_energy_test}
\alias{perform_energy_test}
\title{Perform energy test for multivariate shift}
\usage{
perform_energy_test(X_rct, X_obs, alpha, R)
}
\arguments{
\item{X_rct}{RCT covariates}

\item{X_obs}{Observational covariates}

\item{alpha}{Significance level}

\item{R}{Number of permutations}
}
\value{
Energy test results
}
\description{
Perform energy test for multivariate shift
}
