% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{perform_ks_tests}
\alias{perform_ks_tests}
\title{Perform KS tests for each covariate}
\usage{
perform_ks_tests(X_rct, X_obs, alpha)
}
\arguments{
\item{X_rct}{RCT covariates}

\item{X_obs}{Observational covariates}

\item{alpha}{Significance level}
}
\value{
KS test results
}
\description{
Perform KS tests for each covariate
}
