% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Plot bootstrap distribution}
\usage{
plot_bootstrap(x, ...)
}
\arguments{
\item{x}{unconf_test object}

\item{...}{Additional arguments}
}
\description{
Plot bootstrap distribution
}
