% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{print.batch_unconf_test}
\alias{print.batch_unconf_test}
\title{Print method for batch results}
\usage{
\method{print}{batch_unconf_test}(x, ...)
}
\arguments{
\item{x}{batch_unconf_test object}

\item{...}{Additional arguments}
}
\description{
Print method for batch results
}
\seealso{
\code{\link[=batch_analysis]{batch_analysis()}} for creating batch results.
}
