% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_diagnostic_summary}
\alias{print_diagnostic_summary}
\title{Print diagnostic summary}
\usage{
print_diagnostic_summary(diag_summary)
}
\arguments{
\item{diag_summary}{Diagnostic summary object}
}
\description{
Print diagnostic summary
}
