% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_and_plotting.R
\name{print_sensitivity_summary}
\alias{print_sensitivity_summary}
\title{Print sensitivity summary}
\usage{
print_sensitivity_summary(sensitivity)
}
\arguments{
\item{sensitivity}{Sensitivity analysis object}
}
\description{
Print sensitivity summary
}
