% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{safe_extract}
\alias{safe_extract}
\title{Safely extract element from list}
\usage{
safe_extract(x, element, default = NA)
}
\arguments{
\item{x}{List or vector}

\item{element}{Element name or index}

\item{default}{Default value if element doesn't exist}
}
\value{
Element value or default
}
\description{
Safely extract element from list
}
