% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_and_diagnostics.R
\name{single_bootstrap_replicate}
\alias{single_bootstrap_replicate}
\title{Single bootstrap replicate}
\usage{
single_bootstrap_replicate(
  parsed_data,
  estimator,
  effect_measure,
  family_y,
  transport_applied
)
}
\arguments{
\item{parsed_data}{Parsed data}

\item{estimator}{Estimator}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}

\item{transport_applied}{Whether transport was applied}
}
\value{
Bootstrap difference estimate
}
\description{
Single bootstrap replicate
}
