% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Input validation function}
\usage{
validate_inputs(
  data_rct,
  data_obs,
  formula,
  estimator,
  effect_measure,
  family_y,
  alpha,
  trim
)
}
\arguments{
\item{data_rct}{RCT data}

\item{data_obs}{Observational data}

\item{formula}{Model formula}

\item{estimator}{Chosen estimator}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}

\item{alpha}{Alpha level}

\item{trim}{Trimming bounds}
}
\value{
List with validation results
}
\description{
Input validation function
}
