% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{wtd.var}
\alias{wtd.var}
\title{Weighted variance}
\usage{
wtd.var(x, weights = NULL)
}
\arguments{
\item{x}{Vector}

\item{weights}{Weights}
}
\value{
Weighted variance
}
\description{
Weighted variance
}
