% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{apply_weight_trimming}
\alias{apply_weight_trimming}
\title{Apply weight trimming}
\usage{
apply_weight_trimming(weights, trim)
}
\arguments{
\item{weights}{Weight vector}

\item{trim}{Trimming bounds}
}
\value{
Indices to keep
}
\description{
Apply weight trimming
}
