% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{compute_aipw_influence_function}
\alias{compute_aipw_influence_function}
\title{Compute AIPW influence function}
\usage{
compute_aipw_influence_function(
  pseudo_outcomes,
  weights,
  mu1,
  mu0,
  effect_measure
)
}
\arguments{
\item{pseudo_outcomes}{Pseudo-outcomes}

\item{weights}{Sample weights}

\item{mu1}{Mean under treatment}

\item{mu0}{Mean under control}

\item{effect_measure}{Effect measure}
}
\value{
Influence function
}
\description{
Compute AIPW influence function
}
