% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{compute_aipw_pseudo_outcomes}
\alias{compute_aipw_pseudo_outcomes}
\title{Compute AIPW pseudo-outcomes}
\usage{
compute_aipw_pseudo_outcomes(Y, A, e, mu1_hat, mu0_hat, weights, stabilize)
}
\arguments{
\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{e}{Propensity scores}

\item{mu1_hat}{Outcome predictions under treatment}

\item{mu0_hat}{Outcome predictions under control}

\item{weights}{Sample weights}

\item{stabilize}{Whether to stabilize}
}
\value{
Pseudo-outcomes
}
\description{
Compute AIPW pseudo-outcomes
}
