% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{compute_effect_measure}
\alias{compute_effect_measure}
\title{Compute effect measure from potential outcome means}
\usage{
compute_effect_measure(mu1, mu0, effect_measure)
}
\arguments{
\item{mu1}{Mean under treatment}

\item{mu0}{Mean under control}

\item{effect_measure}{Effect measure}
}
\value{
Effect estimate
}
\description{
Compute effect measure from potential outcome means
}
