% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{compute_effects}
\alias{compute_effects}
\title{Compute effects using specified estimator}
\usage{
compute_effects(
  parsed_data,
  estimator,
  effect_measure,
  family_y,
  w_rct,
  w_obs,
  stabilize,
  trim
)
}
\arguments{
\item{parsed_data}{Parsed data object}

\item{estimator}{Estimation method}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}

\item{w_rct}{RCT weights}

\item{w_obs}{Observational weights}

\item{stabilize}{Stabilize weights}

\item{trim}{Trimming bounds}
}
\value{
Effects object
}
\description{
Compute effects using specified estimator
}
