% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{compute_ipw_influence_function}
\alias{compute_ipw_influence_function}
\title{Compute IPW influence function}
\usage{
compute_ipw_influence_function(Y, A, e, weights, mu1, mu0, effect_measure)
}
\arguments{
\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{e}{Propensity scores}

\item{weights}{IPW weights}

\item{mu1}{Mean under treatment}

\item{mu0}{Mean under control}

\item{effect_measure}{Effect measure}
}
\value{
Influence function
}
\description{
Compute IPW influence function
}
