% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{compute_ipw_weights}
\alias{compute_ipw_weights}
\title{Compute IPW weights}
\usage{
compute_ipw_weights(A, e, weights, stabilize)
}
\arguments{
\item{A}{Treatment vector}

\item{e}{Propensity scores}

\item{weights}{Sample weights}

\item{stabilize}{Whether to stabilize}
}
\value{
IPW weights
}
\description{
Compute IPW weights
}
