% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{compute_standardized_mean_differences}
\alias{compute_standardized_mean_differences}
\title{Compute standardized mean differences}
\usage{
compute_standardized_mean_differences(X_rct, X_obs)
}
\arguments{
\item{X_rct}{RCT covariates}

\item{X_obs}{Observational covariates}
}
\value{
SMD results
}
\description{
Compute standardized mean differences
}
