% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{compute_transport_weights}
\alias{compute_transport_weights}
\title{Compute transport weights}
\usage{
compute_transport_weights(X_rct, X_obs, apply_transport)
}
\arguments{
\item{X_rct}{RCT covariate matrix}

\item{X_obs}{Observational covariate matrix}

\item{apply_transport}{Whether to apply transport weighting}
}
\value{
List with weights
}
\description{
Compute transport weights
}
