% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{create_diagnostic_report}
\alias{create_diagnostic_report}
\title{Create diagnostic report}
\usage{
create_diagnostic_report(
  x,
  file = "unconfoundedness_report.html",
  format = c("html", "pdf")
)
}
\arguments{
\item{x}{unconf_test object}

\item{file}{Output file path (HTML or PDF)}

\item{format}{Output format ("html" or "pdf")}
}
\value{
File path of created report
}
\description{
Create diagnostic report
}
