% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimate_aipw}
\alias{estimate_aipw}
\title{AIPW (Doubly Robust) Estimator}
\usage{
estimate_aipw(Y, A, X, weights, effect_measure, family_y, stabilize, trim)
}
\arguments{
\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{X}{Covariate matrix}

\item{weights}{Sample weights}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}

\item{stabilize}{Stabilize weights}

\item{trim}{Trimming bounds}
}
\value{
AIPW estimate
}
\description{
AIPW (Doubly Robust) Estimator
}
