% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{estimate_tmle}
\alias{estimate_tmle}
\title{TMLE Estimator}
\usage{
estimate_tmle(Y, A, X, weights, effect_measure, family_y)
}
\arguments{
\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{X}{Covariate matrix}

\item{weights}{Sample weights}

\item{effect_measure}{Effect measure}

\item{family_y}{Outcome family}
}
\value{
TMLE estimate
}
\description{
TMLE Estimator
}
