% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_and_helpers.R
\name{export_to_csv}
\alias{export_to_csv}
\title{Export results to CSV}
\usage{
export_to_csv(x, file)
}
\arguments{
\item{x}{unconf_test object or list of results}

\item{file}{Output CSV file path}
}
\value{
Invisibly returns the data frame
}
\description{
Export results to CSV
}
