% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_outcome_models}
\alias{fit_outcome_models}
\title{Fit outcome regression models}
\usage{
fit_outcome_models(Y, A, X, weights, family_y)
}
\arguments{
\item{Y}{Outcome vector}

\item{A}{Treatment vector}

\item{X}{Covariate matrix}

\item{weights}{Sample weights}

\item{family_y}{Outcome family}
}
\value{
Outcome model fits
}
\description{
Fit outcome regression models
}
