% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconfoundedness_test.R
\name{fit_propensity_model}
\alias{fit_propensity_model}
\title{Fit propensity score model}
\usage{
fit_propensity_model(A, X, weights)
}
\arguments{
\item{A}{Treatment vector}

\item{X}{Covariate matrix}

\item{weights}{Sample weights}
}
\value{
Propensity score fit
}
\description{
Fit propensity score model
}
