% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{format_p_value}
\alias{format_p_value}
\title{Format p-value for display}
\usage{
format_p_value(p)
}
\arguments{
\item{p}{P-value}
}
\value{
Formatted p-value string
}
\description{
Format p-value for display
}
