% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.unconf_test}
\alias{plot.unconf_test}
\title{Plot method for unconf_test objects}
\usage{
\method{plot}{unconf_test}(x, type = c("effects", "overlap", "balance", "weights", "bootstrap"), ...)
}
\arguments{
\item{x}{An object of class \code{unconf_test}}

\item{type}{Type of plot: "effects", "overlap", "balance", "weights", "bootstrap"}

\item{...}{Additional plotting arguments}
}
\value{
Invisibly returns the plot object (if applicable)
}
\description{
Plot method for unconf_test objects
}
