% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot covariate balance}
\usage{
plot_balance(x, ...)
}
\arguments{
\item{x}{unconf_test object}

\item{...}{Additional arguments}
}
\description{
Plot covariate balance
}
