% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_and_plotting.R
\name{plot_effects}
\alias{plot_effects}
\title{Plot effect estimates with confidence intervals}
\usage{
plot_effects(x, ...)
}
\arguments{
\item{x}{unconf_test object}

\item{...}{Additional arguments}
}
\description{
Plot effect estimates with confidence intervals
}
