% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_and_plotting.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot propensity score overlap}
\usage{
plot_overlap(x, ...)
}
\arguments{
\item{x}{unconf_test object}

\item{...}{Additional arguments}
}
\description{
Plot propensity score overlap
}
