% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot transport weights distribution}
\usage{
plot_weights(x, ...)
}
\arguments{
\item{x}{unconf_test object}

\item{...}{Additional arguments}
}
\description{
Plot transport weights distribution
}
