% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.unconf_test}
\alias{print.summary.unconf_test}
\title{Print method for summary.unconf_test}
\usage{
\method{print}{summary.unconf_test}(x, ...)
}
\arguments{
\item{x}{Summary object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for summary.unconf_test
}
