% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.unconf_test}
\alias{print.unconf_test}
\title{Enhanced print method for unconf_test objects}
\usage{
\method{print}{unconf_test}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{unconf_test}}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Enhanced print method for unconf_test objects
}
